backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.3
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.19
torch==2.9.0
torchvision==0.24.0
Pillow==12.0.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
fastapi-mail==1.4.1
aiosmtplib==3.0.1